ui: Rename DatasetSliceTrack to SliceTrack
