+++ ui/src/plugins/com.android.AndroidDesktopMode/index.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {DatasetSliceTrack} from '../../components/tracks/dataset_slice_track';
+import {SliceTrack} from '../../components/tracks/slice_track';
 import {PerfettoPlugin} from '../../public/plugin';
 import {Trace} from '../../public/trace';
 import {TrackNode} from '../../public/workspace';
@@ -30,7 +30,7 @@ export default class implements PerfettoPlugin {
 
     ctx.tracks.registerTrack({
       uri: TRACK_URI,
-      renderer: await DatasetSliceTrack.createMaterialized({
+      renderer: await SliceTrack.createMaterialized({
         trace: ctx,
         uri: TRACK_URI,
         dataset: new SourceDataset({
