+++ ui/src/plugins/com.android.AndroidStartup/index.ts
@@ -15,7 +15,7 @@
 import {LONG, NUM, STR} from '../../trace_processor/query_result';
 import {Trace} from '../../public/trace';
 import {PerfettoPlugin} from '../../public/plugin';
-import {DatasetSliceTrack} from '../../components/tracks/dataset_slice_track';
+import {SliceTrack} from '../../components/tracks/slice_track';
 import {SourceDataset} from '../../trace_processor/dataset';
 import {TrackNode} from '../../public/workspace';
 import {optimizationsTrack} from './optimizations';
@@ -41,7 +41,7 @@ export default class implements PerfettoPlugin {
     const startupTrackUri = `/android_startups`;
     ctx.tracks.registerTrack({
       uri: startupTrackUri,
-      renderer: await DatasetSliceTrack.createMaterialized({
+      renderer: await SliceTrack.createMaterialized({
         trace: ctx,
         uri: startupTrackUri,
         dataset: new SourceDataset({
@@ -74,7 +74,7 @@ export default class implements PerfettoPlugin {
     const breakdownTrackUri = '/android_startups_breakdown';
     ctx.tracks.registerTrack({
       uri: breakdownTrackUri,
-      renderer: await DatasetSliceTrack.createMaterialized({
+      renderer: await SliceTrack.createMaterialized({
         trace: ctx,
         uri: breakdownTrackUri,
         dataset: new SourceDataset({
