+++ ui/src/plugins/com.android.AndroidStartup/optimizations.ts
@@ -15,7 +15,7 @@
 import {Trace} from '../../public/trace';
 import {STR, LONG, NUM} from '../../trace_processor/query_result';
 import {TrackNode} from '../../public/workspace';
-import {DatasetSliceTrack} from '../../components/tracks/dataset_slice_track';
+import {SliceTrack} from '../../components/tracks/slice_track';
 import {SourceDataset} from '../../trace_processor/dataset';
 import {DebugSliceTrackDetailsPanel} from '../../components/tracks/debug_slice_track_details_panel';
 
@@ -110,7 +110,7 @@ export async function optimizationsTrack(
   const tableName = `_startup_optimization_slices`;
   trace.tracks.registerTrack({
     uri,
-    renderer: DatasetSliceTrack.create({
+    renderer: SliceTrack.create({
       trace: trace,
       uri,
       dataset: new SourceDataset({
@@ -142,7 +142,7 @@ function classLoadingTrack(
   const uri = `/android_startups/${startup.id}/classloading`;
   trace.tracks.registerTrack({
     uri,
-    renderer: DatasetSliceTrack.create({
+    renderer: SliceTrack.create({
       trace,
       uri,
       dataset: new SourceDataset({
