+++ ui/src/plugins/dev.perfetto.HeapProfile/heap_profile_track.ts
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 import {Time} from '../../base/time';
-import {DatasetSliceTrack} from '../../components/tracks/dataset_slice_track';
+import {SliceTrack} from '../../components/tracks/slice_track';
 import {Trace} from '../../public/trace';
 import {SourceDataset} from '../../trace_processor/dataset';
 import {LONG, NUM, STR} from '../../trace_processor/query_result';
@@ -29,7 +29,7 @@ export function createHeapProfileTrack(
   upid: number,
   heapProfileIsIncomplete: boolean,
 ) {
-  return DatasetSliceTrack.create({
+  return SliceTrack.create({
     trace,
     uri,
     dataset: new SourceDataset({
