+++ ui/src/plugins/dev.perfetto.LinuxPerf/perf_samples_profile_track.ts
@@ -24,7 +24,7 @@ import {Timestamp} from '../../components/widgets/timestamp';
 import {Time, time} from '../../base/time';
 import {Flamegraph, FLAMEGRAPH_STATE_SCHEMA} from '../../widgets/flamegraph';
 import {Trace} from '../../public/trace';
-import {DatasetSliceTrack} from '../../components/tracks/dataset_slice_track';
+import {SliceTrack} from '../../components/tracks/slice_track';
 import {SourceDataset} from '../../trace_processor/dataset';
 import {Stack} from '../../widgets/stack';
 
@@ -35,7 +35,7 @@ export function createProcessPerfSamplesProfileTrack(
   uri: string,
   upid: number,
 ) {
-  return DatasetSliceTrack.create({
+  return SliceTrack.create({
     trace,
     uri,
     dataset: new SourceDataset({
@@ -119,7 +119,7 @@ export function createThreadPerfSamplesProfileTrack(
   uri: string,
   utid: number,
 ) {
-  return DatasetSliceTrack.create({
+  return SliceTrack.create({
     trace,
     uri,
     dataset: new SourceDataset({
