+++ ui/src/plugins/org.chromium.ChromeScrollJank/flat_colored_duration_track.ts
@@ -23,7 +23,7 @@ import {DurationWidget} from '../../components/widgets/duration';
 import {Timestamp} from '../../components/widgets/timestamp';
 import {LONG, NUM, STR} from '../../trace_processor/query_result';
 import {SourceDataset} from '../../trace_processor/dataset';
-import {DatasetSliceTrack} from '../../components/tracks/dataset_slice_track';
+import {SliceTrack} from '../../components/tracks/slice_track';
 import {Time} from '../../base/time';
 
 /*
@@ -36,7 +36,7 @@ export function createFlatColoredDurationTrack(
   uri: string,
   sqlSrc: string,
 ) {
-  return DatasetSliceTrack.create({
+  return SliceTrack.create({
     trace,
     uri,
     dataset: new SourceDataset({
